feat: create UserDao interface and let UserDaoImpl implement UserDao interface
