+++ slipp/src/main/java/slipp/ManualHandlerMapping.java
@@ -7,7 +7,7 @@ import nextstep.mvc.asis.ForwardController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import slipp.controller.*;
-import slipp.dao.UserDao;
+import slipp.dao.UserDaoImpl;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.HashMap;
@@ -21,11 +21,11 @@ public class ManualHandlerMapping implements HandlerMapping {
         mappings.put("/", new HomeController());
         mappings.put("/users/form", new ForwardController("/user/form.jsp"));
         mappings.put("/users/loginForm", new ForwardController("/user/login.jsp"));
-        mappings.put("/users/login", new LoginController(new UserDao()));
+        mappings.put("/users/login", new LoginController(new UserDaoImpl()));
         mappings.put("/users/profile", new ProfileController());
         mappings.put("/users/logout", new LogoutController());
         mappings.put("/users/updateForm", new UpdateFormUserController());
-        mappings.put("/users/update", new UpdateUserController(new UserDao()));
+        mappings.put("/users/update", new UpdateUserController(new UserDaoImpl()));
 
         logger.info("Initialized Request Mapping!");
         mappings.keySet().forEach(path -> {
