+++ slipp/src/main/java/slipp/controller/ApiUserController.java
@@ -1,7 +1,6 @@
 package slipp.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import nextstep.jdbc.DBConnection;
 import nextstep.mvc.JsonView;
 import nextstep.mvc.ModelAndView;
 import nextstep.web.annotation.Controller;
@@ -11,6 +10,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import slipp.dao.UserDao;
+import slipp.dao.UserDaoImpl;
 import slipp.domain.User;
 import slipp.dto.UserCreatedDto;
 import slipp.dto.UserUpdatedDto;
@@ -27,7 +27,7 @@ public class ApiUserController {
 
     public ApiUserController() {
         this.objectMapper = new ObjectMapper();
-        this.userDao = new UserDao();
+        this.userDao = new UserDaoImpl();
     }
 
     @RequestMapping(value = "/api/users", method = RequestMethod.POST)
